src/requirements.txt
line-bot-sdk
requests
urllib3<2